New for loop instead of If statements.
I changed the logic to a couple of for loops instead of if statements.  I makes it a bit easier to read.  There seems to be a small flaw in my logic I am not able to currently see.  But this for loop should make it much easier to display states later on.